+++ src/main/java/org/motechproject/newebodac/service/VisitTypeService.java
@@ -2,10 +2,13 @@ package org.motechproject.newebodac.service;
 
 import java.util.List;
 import java.util.UUID;
+import org.motechproject.newebodac.domain.Visit;
 import org.motechproject.newebodac.domain.VisitType;
 import org.motechproject.newebodac.dto.VisitTypeDto;
 import org.motechproject.newebodac.exception.EntityNotFoundException;
+import org.motechproject.newebodac.exception.RelationViolationException;
 import org.motechproject.newebodac.mapper.VisitTypeMapper;
+import org.motechproject.newebodac.repository.VisitRepository;
 import org.motechproject.newebodac.repository.VisitTypeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,6 +20,9 @@ public class VisitTypeService {
 
   @Autowired
   private VisitTypeRepository visitTypeRepository;
+
+  @Autowired
+  private VisitRepository visitRepository;
 
   public List<VisitTypeDto> getAll() {
     return MAPPER.toDtos(visitTypeRepository.findAll());
@@ -31,12 +37,32 @@ public class VisitTypeService {
   }
 
   /**
+   * Updates VisitType from it's dto.
+   * @param id id of updated VisitType.
+   * @param visitTypeDto dto of updated VisitType.
+   * @return Dto of updated VisitType.
+   */
+  public VisitTypeDto update(UUID id, VisitTypeDto visitTypeDto) {
+    VisitType visitType = visitTypeRepository.findById(id)
+        .orElseThrow(() -> new EntityNotFoundException("Visit type with id: {0} not found",
+            id.toString()));
+    MAPPER.update(visitTypeDto, visitType);
+    return MAPPER.toDto(visitTypeRepository.save(visitType));
+  }
+
+  /**
    * Deletes visit type with given id.
    * @param id ID of visit type to delete.
    */
   public void delete(UUID id) {
     VisitType visitType = visitTypeRepository.findById(id).orElseThrow(() ->
         new EntityNotFoundException("Visit Type with id: {0} not found", id.toString()));
+
+    List<Visit> visits = visitRepository.getByType(visitType);
+    if (!visits.isEmpty()) {
+      throw new RelationViolationException("Can not delete visit type with visits in it!");
+    }
+
     visitTypeRepository.delete(visitType);
   }
 }
